#409 Change definitions of syntax sugars
