+++ neclib/core/data_type/__init__.py
@@ -3,5 +3,6 @@
 from .ordinal import Ordinal
 from .parameters import Parameters
 from .rich_parameters import RichParameters
+from .value_range import ValueRange
 
-__all__ = ["Ordinal", "Parameters", "RichParameters"]
+__all__ = ["Ordinal", "Parameters", "RichParameters", "ValueRange"]
